+++ src/Umbraco.Core/Models/TemporaryFileUpload/CreateTemporaryFileModel.cs
@@ -1,0 +1,5 @@
+﻿namespace Umbraco.Cms.Core.Models.TemporaryFile;
+
+public class CreateTemporaryFileModel : TemporaryFileModelBase
+{
+}
