+++ src/Umbraco.Core/Models/TemporaryFileUpload/TemporaryFileModel.cs
@@ -4,16 +4,3 @@ public class TemporaryFileModel : TemporaryFileModelBase
 {
     public required DateTime AvailableUntil { get; set; }
 }
-
-public class CreateTemporaryFileModel : TemporaryFileModelBase
-{
-}
-
-public abstract class TemporaryFileModelBase
-{
-    public required string FileName { get; set; }
-
-    public Guid Key { get; set; }
-
-    public Func<Stream> OpenReadStream { get; set; } = () => Stream.Null;
-}
